ofActionPayload helper is provided
